Fix compiler warnings on FreeBSD/alpha.
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@5091 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/cpus/cpu_alpha_instr.cc
@@ -25,7 +25,7 @@
  *  SUCH DAMAGE.
  *
  *
- *  $Id: cpu_alpha_instr.cc,v 1.1 2007-11-17 11:15:30 debug Exp $
+ *  $Id: cpu_alpha_instr.cc,v 1.2 2007-11-17 11:28:40 debug Exp $
  *
  *  Alpha instructions.
  *
@@ -466,7 +466,7 @@ X(cvttq_c)
 {
 	struct ieee_float_value fb;
 	ieee_interpret_float_value(reg(ic->arg[2]), &fb, IEEE_FMT_D);
-	reg(ic->arg[0]) = fb.nan? 0 : fb.f;
+	reg(ic->arg[0]) = (int64_t) ( fb.nan? 0 : fb.f );
 }
